+++ mpp/src/MaterialStream.cpp
@@ -69,7 +69,7 @@ namespace mpp
 
 	uint32_t MaterialStream::createQualitySetting(string const& name)
 	{
-		auto qualityId = mQualitySettings.size();
+		auto qualityId = (uint32_t)mQualitySettings.size();
 		mQualityNames[name] = qualityId;
 
 		mQualitySettings.push_back(QualitySetting());
