+++ mpp/src/PostEffectStream.cpp
@@ -30,7 +30,7 @@ namespace mpp
 
 	uint32_t PostEffectStream::createQualitySetting(string const& name)
 	{
-		auto qualityId = mQualitySettings.size();
+		auto qualityId = (uint32_t)mQualitySettings.size();
 		mQualityNames[name] = qualityId;
 
 		mQualitySettings.push_back(QualitySetting());
