+++ mpp/src/StringStream.cpp
@@ -16,7 +16,7 @@ namespace mpp
 
 	uint32_t StringStream::createQualitySetting(string const& name)
 	{
-		auto qualityId = mQualitySettings.size();
+		auto qualityId = (uint32_t)mQualitySettings.size();
 		mQualityNames[name] = qualityId;
 
 		mQualitySettings.push_back(QualitySetting());
