+++ mpp/src/TextureStream.cpp
@@ -144,7 +144,7 @@ namespace mpp
 
 	uint32_t TextureStream::createQualitySetting(string const& name)
 	{
-		auto qualityId = mQualitySettings.size();
+		auto qualityId = (uint32_t)mQualitySettings.size();
 		mQualityNames[name] = qualityId;
 
 		mQualitySettings.push_back(QualitySetting());
